Use instanced configuration from AbstractFileSystem
