+++ core/client/hdfs/src/main/java/alluxio/hadoop/AbstractFileSystem.java
@@ -14,8 +14,10 @@ package alluxio.hadoop;
 import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toMap;
 
+import alluxio.AlluxioConfiguration;
 import alluxio.AlluxioURI;
 import alluxio.Configuration;
+import alluxio.InstancedConfiguration;
 import alluxio.PropertyKey;
 import alluxio.client.block.AlluxioBlockStore;
 import alluxio.client.block.BlockWorkerInfo;
@@ -495,7 +497,7 @@ abstract class AbstractFileSystem extends org.apache.hadoop.fs.FileSystem {
     // Load Alluxio configuration if any and merge to the one in Alluxio file system. These
     // modifications to ClientContext are global, affecting all Alluxio clients in this JVM.
     // We assume here that all clients use the same configuration.
-    HadoopConfigurationUtils.mergeHadoopConfiguration(conf);
+    HadoopConfigurationUtils.mergeHadoopConfiguration(conf, Configuration.global());
     Configuration.set(PropertyKey.ZOOKEEPER_ENABLED, isZookeeperMode());
     // When using zookeeper we get the leader master address from the alluxio.zookeeper.address
     // configuration property, so the user doesn't need to specify the authority.
@@ -532,21 +534,11 @@ abstract class AbstractFileSystem extends org.apache.hadoop.fs.FileSystem {
   private boolean connectDetailsMatch(URI uri, org.apache.hadoop.conf.Configuration conf) {
     // Create the master inquire client that we would have after merging the hadoop conf into
     // Alluxio Configuration.
-    MasterInquireClient.Factory.Config inquireClientConf =
-        MasterInquireClient.Factory.Config.defaults();
-    inquireClientConf.setZookeeperEnabled(conf.getBoolean(PropertyKey.ZOOKEEPER_ENABLED.getName(),
-        inquireClientConf.isZookeeperEnabled()));
-    inquireClientConf.setZookeeperAddress(
-        conf.get(PropertyKey.ZOOKEEPER_ADDRESS.getName(), inquireClientConf.getZookeeperAddress()));
-    inquireClientConf.setElectionPath(conf.get(PropertyKey.ZOOKEEPER_ELECTION_PATH.getName(),
-        inquireClientConf.getElectionPath()));
-    inquireClientConf.setLeaderPath(
-        conf.get(PropertyKey.ZOOKEEPER_ELECTION_PATH.getName(), inquireClientConf.getLeaderPath()));
-    inquireClientConf.setConnectHost(mUri.getHost());
-    inquireClientConf.setConnectPort(mUri.getPort());
-    MasterInquireClient configClient = MasterInquireClient.Factory.create(inquireClientConf);
-    MasterInquireClient contextClient = FileSystemContext.get().getMasterInquireClient();
-    return configClient.equals(contextClient);
+    AlluxioConfiguration alluxioConf = new InstancedConfiguration(Configuration.global());
+    HadoopConfigurationUtils.mergeHadoopConfiguration(conf, alluxioConf);
+    MasterInquireClient configClient = MasterInquireClient.Factory.create(alluxioConf);
+
+    return configClient.equals(FileSystemContext.get().getMasterInquireClient());
   }
 
   /**
