+++ core/client/hdfs/src/main/java/alluxio/hadoop/HadoopConfigurationUtils.java
@@ -11,6 +11,7 @@
 
 package alluxio.hadoop;
 
+import alluxio.AlluxioConfiguration;
 import alluxio.Configuration;
 import alluxio.PropertyKey;
 import alluxio.conf.Source;
@@ -54,11 +55,13 @@ public final class HadoopConfigurationUtils {
   }
 
   /**
-   * Merges Hadoop {@link org.apache.hadoop.conf.Configuration} into the Alluxio configuration.
+   * Merges Hadoop {@link org.apache.hadoop.conf.Configuration} into the Alluxio properties.
    *
    * @param source the {@link org.apache.hadoop.conf.Configuration} to merge
+   * @param alluxioConfiguration the Alluxio configuration to merge to
    */
-  public static void mergeHadoopConfiguration(org.apache.hadoop.conf.Configuration source) {
+  public static void mergeHadoopConfiguration(org.apache.hadoop.conf.Configuration source,
+      AlluxioConfiguration alluxioConfiguration) {
     // Load Alluxio configuration if any and merge to the one in Alluxio file system
     // Push Alluxio configuration to the Job configuration
     Properties alluxioConfProperties = new Properties();
@@ -72,7 +75,7 @@ public final class HadoopConfigurationUtils {
     LOG.info("Loading Alluxio properties from Hadoop configuration: {}", alluxioConfProperties);
     // Merge the relevant Hadoop configuration into Alluxio's configuration.
     // TODO(jiri): support multiple client configurations (ALLUXIO-2034)
-    Configuration.merge(alluxioConfProperties, Source.RUNTIME);
-    Configuration.validate();
+    alluxioConfiguration.merge(alluxioConfProperties, Source.RUNTIME);
+    alluxioConfiguration.validate();
   }
 }
