+++ core/common/src/main/java/alluxio/master/MasterInquireClient.java
@@ -11,6 +11,7 @@
 
 package alluxio.master;
 
+import alluxio.AlluxioConfiguration;
 import alluxio.Configuration;
 import alluxio.PropertyKey;
 import alluxio.exception.status.UnavailableException;
@@ -51,156 +52,25 @@ public interface MasterInquireClient {
      * @return a master inquire client
      */
     public static MasterInquireClient create() {
-      return create(Config.defaults());
+      return create(Configuration.global());
     }
 
     /**
      * @param config configuration for creating the master inquire client
      * @return a master inquire client
      */
-    public static MasterInquireClient create(Config config) {
-      if (config.isZookeeperEnabled()) {
-        return ZkMasterInquireClient.getClient(config.getZookeeperAddress(),
-            config.getElectionPath(), config.getLeaderPath());
+    public static MasterInquireClient create(AlluxioConfiguration config) {
+      if (config.getBoolean(PropertyKey.ZOOKEEPER_ENABLED)) {
+        return ZkMasterInquireClient.getClient(config.get(PropertyKey.ZOOKEEPER_ADDRESS),
+            config.get(PropertyKey.ZOOKEEPER_ELECTION_PATH),
+            config.get(PropertyKey.ZOOKEEPER_LEADER_PATH));
       } else {
         return new SingleMasterInquireClient(
-            new InetSocketAddress(config.getConnectHost(), config.getConnectPort()));
+            NetworkAddressUtils.getConnectAddress(ServiceType.MASTER_RPC, config));
       }
     }
 
-    private Factory() {} // Not intended for instantiation.
-
-    /**
-     * Configuration for building a {@link MasterInquireClient} from a
-     * {@link MasterInquireClient.Factory}.
-     */
-    public static final class Config {
-      // HA connect with Zookeeper.
-      private boolean mZookeeperEnabled;
-      private String mZookeeperAddress;
-      private String mElectionPath;
-      private String mLeaderPath;
-
-      // Non-HA connect.
-      private String mConnectHost;
-      private int mConnectPort;
-
-      // Use Config.defaults() instead.
-      private Config() {}
-
-      /**
-       * @return the default master inquire configuration based on {@link Configuration}
-       */
-      public static Config defaults() {
-        String zkAddress = Configuration.containsKey(PropertyKey.ZOOKEEPER_ADDRESS)
-            ? Configuration.get(PropertyKey.ZOOKEEPER_ADDRESS)
-            : null;
-        return new Config()
-            .setZookeeperEnabled(Configuration.getBoolean(PropertyKey.ZOOKEEPER_ENABLED))
-            .setZookeeperAddress(zkAddress)
-            .setElectionPath(Configuration.get(PropertyKey.ZOOKEEPER_ELECTION_PATH))
-            .setLeaderPath(Configuration.get(PropertyKey.ZOOKEEPER_LEADER_PATH))
-            .setConnectHost(
-                NetworkAddressUtils.getConnectAddress(ServiceType.MASTER_RPC).getHostName())
-            .setConnectPort(
-                NetworkAddressUtils.getConnectAddress(ServiceType.MASTER_RPC).getPort());
-      }
-
-      /**
-       * @param zookeeperEnabled whether zookeeper is enabled
-       * @return this
-       */
-      public Config setZookeeperEnabled(boolean zookeeperEnabled) {
-        mZookeeperEnabled = zookeeperEnabled;
-        return this;
-      }
-
-      /**
-       * @param zookeeperAddress zookeeper address
-       * @return this
-       */
-      public Config setZookeeperAddress(String zookeeperAddress) {
-        mZookeeperAddress = zookeeperAddress;
-        return this;
-      }
-
-      /**
-       * @param electionPath election path
-       * @return this
-       */
-      public Config setElectionPath(String electionPath) {
-        mElectionPath = electionPath;
-        return this;
-      }
-
-      /**
-       * @param leaderPath leader path
-       * @return this
-       */
-      public Config setLeaderPath(String leaderPath) {
-        mLeaderPath = leaderPath;
-        return this;
-      }
-
-      /**
-       * @param host master connect host
-       * @return this
-       */
-      public Config setConnectHost(String host) {
-        mConnectHost = host;
-        return this;
-      }
-
-      /**
-       * @param port master connect port
-       * @return this
-       */
-      public Config setConnectPort(int port) {
-        mConnectPort = port;
-        return this;
-      }
-
-      /**
-       * @return whether zookeeper is enabled
-       */
-      public boolean isZookeeperEnabled() {
-        return mZookeeperEnabled;
-      }
-
-      /**
-       * @return the zookeeper address
-       */
-      public String getZookeeperAddress() {
-        return mZookeeperAddress;
-      }
-
-      /**
-       * @return the election path
-       */
-      public String getElectionPath() {
-        return mElectionPath;
-      }
-
-      /**
-       * @return the leader path
-       */
-      public String getLeaderPath() {
-        return mLeaderPath;
-      }
-
-      /**
-       * @return the connect host
-       */
-      public String getConnectHost() {
-        return mConnectHost;
-      }
-
-      /**
-       * @return the connect port
-       */
-      public int getConnectPort() {
-        return mConnectPort;
-      }
-    }
+    private Factory() {
+    } // Not intended for instantiation.
   }
 }
